Create README.md
This project analyzes the relationship between the S&P 500 stock index and three independent variables: the U.S. 10-Year Treasury Yield, Crude Oil Prices, and the U.S. Dollar Index. Using daily data from April 2020 to March 2023, a multiple regression analysis was performed to evaluate the predictive ability of these variables in explaining the movement of the S&P 500 index. The regression model shows a strong fit, with an R-squared value of 0.72, suggesting that approximately 72% of the variation in the S&P 500 index can be explained by the chosen independent variables. The individual significance of each variable was assessed, with all showing statistically significant relationships with the S&P 500.